feat(context): Supporting auth props
