+++ module/rest-session/__index__.ts
@@ -2,4 +2,5 @@ import type { } from './src/trv';
 export * from './src/config';
 export * from './src/interceptor';
 export * from './src/service';
+export * from './src/model';
 export * from './src/session';
