+++ module/rest-session/src/model.ts
@@ -1,0 +1,18 @@
+import { Model, ExpiresAt } from '@travetto/model';
+import { Text } from '@travetto/schema';
+
+/**
+ * Session model service identifier
+ */
+export const SessionModelSymbol = Symbol.for('@travetto/rest-session:model');
+
+@Model({ autoCreate: false })
+export class SessionEntry {
+  id: string;
+  @Text()
+  data: string;
+  @ExpiresAt()
+  expiresAt?: Date;
+  issuedAt: Date;
+  maxAge?: number;
+}
