+++ module/rest/src/interceptor/context.ts
@@ -11,17 +11,13 @@ import { BodyParseInterceptor } from './body-parse';
 @Config('rest.context')
 class RestAsyncContextConfig extends ManagedInterceptorConfig { }
 
-const REQUEST_SYMBOL = Symbol.for('@travetto/rest:request');
-const RESPONSE_SYMBOL = Symbol.for('@travetto/rest:response');
-
 /**
  * Enables access to contextual data when running in a rest application
  */
 @Injectable()
 export class AsyncContextInterceptor implements RestInterceptor {
 
-  #reqProp: AsyncContextProp<Request>;
-  #resProp: AsyncContextProp<Response>;
+  #ctxProp: AsyncContextProp<FilterContext>;
 
   dependsOn = [BodyParseInterceptor];
 
@@ -32,23 +28,21 @@ export class AsyncContextInterceptor implements RestInterceptor {
   config: RestAsyncContextConfig;
 
   postConstruct(): void {
-    this.#reqProp = this.context.prop(REQUEST_SYMBOL);
-    this.#resProp = this.context.prop(RESPONSE_SYMBOL);
+    this.#ctxProp = this.context.prop();
   }
 
   intercept(ctx: FilterContext, next: FilterNext): Promise<unknown> {
     return this.context.run(() => {
-      this.#reqProp.set(ctx.req);
-      this.#resProp.set(ctx.res);
+      this.#ctxProp.set(ctx);
       return next();
     });
   }
 
   get request(): Request {
-    return this.#reqProp.get()!;
+    return this.#ctxProp.get()?.req!;
   }
 
   get response(): Response {
-    return this.#resProp.get()!;
+    return this.#ctxProp.get()?.res!;
   }
 }
